# Use thread local `MessageDigest`s for hashing functions