There are still several codepath pointing to sha256 hash functions. As per my previous ssz optimizations I saw that instantiating every time a new `MessageDigest` is quite expensive. `MessageDigest` reset themself everytime `digest()` is called, so they are reusable.

I'd like to see if there are noticable benefits with this approach.

## Documentation

- [ ] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [ ] I thought about adding a changelog entry, and added one if I deemed necessary.