@StefanBratanov 2023-03-24
> `Old`
> 
> ```
> Benchmark                             Mode  Cnt        Score        Error  Units
> Sha256Benchmark.pureKeccak256        thrpt   50  2866031.990 ±  48965.282  ops/s
> Sha256Benchmark.pureSha256           thrpt   50  2926474.482 ±  31450.630  ops/s
> Sha256Benchmark.sha256of33byteArray  thrpt   50  2857816.612 ±  23948.549  ops/s
> Sha256Benchmark.sha256of33bytes      thrpt   50  2718669.114 ± 119768.117  ops/s
> ```
> 
> `New`
> 
> ```
> Benchmark                             Mode  Cnt        Score        Error  Units
> Sha256Benchmark.pureKeccak256        thrpt   50  3347330.321 ± 141670.181  ops/s
> Sha256Benchmark.pureSha256           thrpt   50  3069017.952 ±  28841.554  ops/s
> Sha256Benchmark.sha256of33byteArray  thrpt   50  3097595.629 ±  32631.950  ops/s
> Sha256Benchmark.sha256of33bytes      thrpt   50  3067835.616 ±  35049.572  ops/s
> ```

Nice. Looks better.