@tbenr 2023-03-24
10milion rounds

old
<img width="1348" alt="image" src="https://user-images.githubusercontent.com/15999009/227557315-a8150cc8-a1f2-4ffa-86a4-d4dab7b8903f.png">

new
<img width="1275" alt="image" src="https://user-images.githubusercontent.com/15999009/227557056-02669eeb-39b9-4b2d-b26f-0ae6ec26c405.png">